+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas.geometry.curves.curve.Curve.from_native`.
 * Added `compas_rhino.geometry.curves.curve.Curve.from_native`.
 * Added `compas_rhino.geometry.curves.nurbs.NurbsCurve.from_native`.
+* Added `compas.datastructures.HashTree` and `compas.datastructures.HashNode`.
 
 ### Changed
 
@@ -165,7 +166,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas.datastructures.CellNetwork.face_plane`
 * Added `compas.datastructures.CellNetwork.cell_volume`
 * Added `compas.datastructures.CellNetwork.cell_neighbors`
-* Added `compas.datastructures.HashTree` and `compas.datastructures.HashNode`.
 
 ### Changed